# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Bingo App API",
        "description": "API for managing bingo rooms and real-time number drawing",
        "contact": {
            "name": "Bingo App",
            "url": "https://github.com/yanajiki/bingoapp"
        },
        "version": "2.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "tags": [
        {
            "name": "Rooms",
            "description": "Bingo room management"
        },
        {
            "name": "Feedback",
            "description": "Feedback form submission"
        }
    ],
    "paths": {
        "/api/v1/room": {
            "post": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Create a bingo room",
                "description": "Creates a new bingo room and returns the creator view, including the creatorHash required for privileged operations.",
                "operationId": "create",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateRoomForm"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Room created successfully",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/RoomDTO"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error \u2014 name is blank or exceeds 255 characters",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "A room with the same name already exists",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/room/lookup": {
            "post": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Lookup rooms by creator hashes",
                "description": "Returns the rooms matching the supplied creator hashes in creator view (including creatorHash). Unknown hashes are silently skipped. A null or empty list yields an empty result. Never returns 404.",
                "operationId": "lookup",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RoomLookupForm"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "List of matching rooms (possibly empty)",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/RoomDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/feedback": {
            "post": {
                "tags": [
                    "Feedback"
                ],
                "summary": "Submit a feedback message",
                "description": "Saves the message and fires an async Discord notification.",
                "operationId": "submit",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FeedbackForm"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Message received",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/FeedbackMessageDTO"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/FeedbackMessageDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/room/{session-code}": {
            "get": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Get a bingo room",
                "description": "Returns the room identified by the given session code. If the X-Creator-Hash header is supplied and matches, the response includes the creatorHash (creator view). Otherwise the creatorHash is omitted (player view).",
                "operationId": "search",
                "parameters": [
                    {
                        "name": "session-code",
                        "in": "path",
                        "description": "Public session code of the room",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "X-Creator-Hash",
                        "in": "header",
                        "description": "Optional creator hash; if valid, returns the creator view including creatorHash",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Room found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/RoomDTO"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Room not found for the given session code",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Delete a bingo room",
                "description": "Permanently deletes the room identified by the given session code. Requires the X-Creator-Hash header to authenticate the request as the room creator.",
                "operationId": "delete",
                "parameters": [
                    {
                        "name": "session-code",
                        "in": "path",
                        "description": "Public session code of the room",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "X-Creator-Hash",
                        "in": "header",
                        "description": "Creator hash for authentication; must match the hash assigned at room creation",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Room deleted successfully"
                    },
                    "400": {
                        "description": "Invalid or missing creator hash",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Room not found for the given session code",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/room/{session-code}/qrcode": {
            "get": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Get QR code for a bingo room",
                "description": "Returns a 250x250 PNG QR code that encodes the join URL for the given room. No authentication required. Returns 404 if the room does not exist.",
                "operationId": "getQrCode",
                "parameters": [
                    {
                        "name": "session-code",
                        "in": "path",
                        "description": "Public session code of the room",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "QR code PNG image",
                        "content": {
                            "image/png": {}
                        }
                    },
                    "404": {
                        "description": "Room not found for the given session code",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/room/{session-code}/players": {
            "get": {
                "tags": [
                    "Rooms"
                ],
                "summary": "List players in a room",
                "description": "Returns all players who have joined the room. Requires creator authentication via X-Creator-Hash header.",
                "operationId": "getPlayers",
                "parameters": [
                    {
                        "name": "session-code",
                        "in": "path",
                        "description": "Public session code of the room",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "X-Creator-Hash",
                        "in": "header",
                        "description": "Creator hash for authentication; must match the hash assigned at room creation",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Player list retrieved successfully",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PlayerDTO"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Room not found or invalid creator hash",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "CreateRoomForm": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 255,
                        "minLength": 0
                    },
                    "description": {
                        "type": "string",
                        "maxLength": 255,
                        "minLength": 0
                    },
                    "drawMode": {
                        "type": "string",
                        "enum": [
                            "MANUAL",
                            "AUTOMATIC"
                        ]
                    }
                }
            },
            "RoomDTO": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "sessionCode": {
                        "type": "string"
                    },
                    "creatorHash": {
                        "type": "string"
                    },
                    "drawnNumbers": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    "drawnLabels": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "drawMode": {
                        "type": "string",
                        "enum": [
                            "MANUAL",
                            "AUTOMATIC"
                        ]
                    }
                }
            },
            "ApiResponse": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    }
                }
            },
            "RoomLookupForm": {
                "type": "object",
                "properties": {
                    "creatorHashes": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
            "FeedbackForm": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 100,
                        "minLength": 0
                    },
                    "email": {
                        "type": "string",
                        "maxLength": 254,
                        "minLength": 0
                    },
                    "phone": {
                        "type": "string",
                        "maxLength": 20,
                        "minLength": 0
                    },
                    "content": {
                        "type": "string",
                        "maxLength": 2000,
                        "minLength": 0
                    }
                }
            },
            "FeedbackMessageDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "content": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "PlayerDTO": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "joinDateTime": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            }
        }
    }
}